% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRTsampleSearch.R
\name{simulate_CRT}
\alias{simulate_CRT}
\title{simulate a smaple CRT study data}
\usage{
simulate_CRT(nt, nc, replacement = TRUE, FUN_clustersize = NULL,
  FUN_Ys = NULL, dataset = NULL, outcome = NULL, clusterID = NULL,
  ...)
}
\arguments{
\item{nt}{the number of cluster in the treatment arm to start the search}

\item{nc}{the number of cluster in the control arm to start the search}

\item{FUN_clustersize}{user-defined function for the distribution of cluster sizes, should take in only one argument: n=the number of clsuters in the study}

\item{FUN_Ys}{user-defined function for the distribution of the two potential outcomes, should take in only one argument: m=the number of individual units in the a clsuter}

\item{dataset}{a dataset from which to generate the study data (no need to provide FUN_clustersize or FUN_Ys), should be individual level data, including columns for clutser ID, outcome of interest, and the variable for stratified randomization if desired.}

\item{outcome}{the name of the outcome variable in the dataset}

\item{clusterID}{the name of the cluster ID in the dataset}

\item{nrep}{the number of replications of the simulation procedure for generating the pseudo study data, default is \code{"1000"})}
}
\value{
a dataframe
}
\description{
data generating mechanism
}
\examples{
\dontrun{
## distribution of cluster sizes
sim_cluster_size=function(N, ...){
  size=round(100*rnorm(N, 0, 1), 0)
  size[size<=0]=1
  return(size)
}
## distribution of the two potential outcomes
sim_potential_outcomes=function(m,...){
  muibar=rnorm(1, 0, 1)
  Y0=rnorm(m, muibar, 10) 
  Y1=Y0 + 1
  re=cbind(Y0, Y1)
  colnames(re)=list("Y0", "Y1")
  return(re)
}

## simulate a CRT study
simulate_CRT(nt=10, nc=10, FUN_clustersize=sim_cluster_size, FUN_Ys=sim_potential_outcomes, ...)
}

}
